requirements.txt
matplotlib==3.2.1
numpy==1.18.4
seaborn==0.10.1
joblib==0.15.1
streamlit==0.59.0
pandas==1.0.3
Pillow==7.1.2
